+++ src/AngleSharp.Css/Dom/Internal/Rules/CssSupportsRule.cs
@@ -3,11 +3,13 @@ namespace AngleSharp.Css.Dom
     using AngleSharp.Css.Parser;
     using AngleSharp.Dom;
     using System;
+    using System.Diagnostics;
     using System.IO;
 
     /// <summary>
     /// Represents an @supports rule.
     /// </summary>
+    [DebuggerDisplay(null, Name = "CssSupportsRule ({ConditionText})")]
     sealed class CssSupportsRule : CssConditionRule, ICssSupportsRule
     {
         #region Fields