+++ src/AngleSharp.Css/Dom/Internal/Rules/CssViewportRule.cs
@@ -1,4 +1,4 @@
-﻿namespace AngleSharp.Css.Dom
+namespace AngleSharp.Css.Dom
 {
     using AngleSharp.Css;
     using System;