+++ src/AngleSharp.Css/Extensions/WindowExtensions.cs
@@ -3,7 +3,9 @@ namespace AngleSharp.Dom
     using AngleSharp.Attributes;
     using AngleSharp.Css;
     using AngleSharp.Css.Dom;
+    using AngleSharp.Css.RenderTree;
     using System;
+    using System.Linq;
 
     /// <summary>
     /// A set of useful extension methods for the Window class.
@@ -36,7 +38,27 @@ namespace AngleSharp.Dom
         [DomName("getPseudoElements")]
         public static ICssPseudoElementList GetPseudoElements(this IWindow window, IElement element, String type = null)
         {
-            throw new NotImplementedException();
+            var validTypes = new[] { "::before", "::after" };
+
+            if (type == null)
+            {
+                // Everything is fine - we take all valid types
+            }
+            else if (validTypes.Contains(type))
+            {
+                validTypes = new[] { type };
+            }
+            else
+            {
+                throw new DomException(DomError.NotSupported);
+            }
+
+            return new CssPseudoElementList(validTypes.Select(pseudoSelector =>
+            {
+                var pseudoElement = element?.Pseudo(pseudoSelector.TrimStart(':'));
+                var style = window.GetComputedStyle(pseudoElement);
+                return new CssPseudoElement(pseudoElement, pseudoSelector, style);
+            }));
         }
 
         /// <summary>
@@ -89,6 +111,21 @@ namespace AngleSharp.Dom
             // Replaces the relative values with absolute ones
             // --> computed
             throw new NotImplementedException();
+        }
+
+        /// <summary>
+        /// Renders the currently available document into a render tree rooted at the returned render node.
+        /// a render tree is essentially the combination of DOM nodes with their CSSOM computed style declarations.
+        ///
+        /// In case no render device is supplied the context's default render device is chosen.
+        /// </summary>
+        /// <param name="window">The window to extend.</param>
+        /// <param name="renderDevice">The device for rendering, if any. </param>
+        /// <returns>The created render node.</returns>
+        public static IRenderNode Render(this IWindow window, IRenderDevice renderDevice = null)
+        {
+            var builder = new RenderTreeBuilder(window, renderDevice);
+            return builder.RenderDocument();
         }
     }
 }