+++ src/AngleSharp.Css/Values/Functions/CssScaleValue.cs
@@ -136,7 +136,7 @@ namespace AngleSharp.Css.Values
         /// Computes the matrix for the given transformation.
         /// </summary>
         /// <returns>The transformation matrix representation.</returns>
-        public TransformMatrix ComputeMatrix() =>
+        public TransformMatrix ComputeMatrix(IRenderDimensions renderDimensions) =>
             new TransformMatrix(_sx, 0f, 0f, 0f, _sy, 0f, 0f, 0f, _sz, 0f, 0f, 0f, 0f, 0f, 0f);
 
         #endregion